requirements-gpu.txt
# Installation rapide pour GPU NVIDIA
# Utilise les index PyTorch officiels pour CUDA 12.1

--index-url https://download.pytorch.org/whl/cu121
torch==2.5.1+cu121
torchvision==0.20.1+cu121  
torchaudio==2.5.1+cu121

--index-url https://pypi.org/simple/
diffusers==0.35.2
transformers==4.57.1
accelerate==1.11.0
pillow==12.0.0
numpy==2.3.4